move workspace related class to monitor lib
