+++ azure-toolkit-libs/azure-toolkit-applicationinsights-lib/pom.xml
@@ -109,5 +109,11 @@
             <groupId>com.microsoft.azure</groupId>
             <artifactId>azure-toolkit-auth-lib</artifactId>
         </dependency>
+        <dependency>
+            <groupId>com.microsoft.azure</groupId>
+            <artifactId>azure-toolkit-monitor-lib</artifactId>
+            <version>0.28.0-SNAPSHOT</version>
+            <scope>compile</scope>
+        </dependency>
     </dependencies>
 </project>
