+++ azure-toolkit-libs/azure-toolkit-applicationinsights-lib/src/main/resources/META-INF/services/com.microsoft.azure.toolkit.lib.AzService
@@ -1,2 +1,1 @@
 com.microsoft.azure.toolkit.lib.applicationinsights.AzureApplicationInsights
-com.microsoft.azure.toolkit.lib.applicationinsights.workspace.AzureLogAnalyticsWorkspace
