+++ azure-toolkit-libs/azure-toolkit-monitor-lib/pom.xml
@@ -1,0 +1,108 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+<parent>
+    <artifactId>azure-toolkit-libs</artifactId>
+    <groupId>com.microsoft.azure</groupId>
+    <version>0.28.0-SNAPSHOT</version>
+</parent>
+<modelVersion>4.0.0</modelVersion>
+
+<artifactId>azure-toolkit-monitor-lib</artifactId>
+<build>
+    <plugins>
+        <plugin>
+            <groupId>org.apache.maven.plugins</groupId>
+            <artifactId>maven-dependency-plugin</artifactId>
+        </plugin>
+        <plugin>
+            <groupId>org.apache.maven.plugins</groupId>
+            <artifactId>maven-source-plugin</artifactId>
+        </plugin>
+        <plugin>
+            <groupId>org.apache.maven.plugins</groupId>
+            <artifactId>maven-javadoc-plugin</artifactId>
+            <configuration>
+                <show>private</show>
+                <failOnError>false</failOnError>
+            </configuration>
+            <executions>
+                <execution>
+                    <id>attach-javadocs</id>
+                    <goals>
+                        <goal>jar</goal>
+                    </goals>
+                    <configuration>
+                        <additionalparam>${javadoc.opts}</additionalparam>
+                    </configuration>
+                </execution>
+            </executions>
+        </plugin>
+        <plugin>
+            <!-- https://mvnrepository.com/artifact/org.codehaus.mojo/aspectj-maven-plugin -->
+            <!-- http://www.quabr.com/62976155/aspectj-maven-plugin-1-11-missing-tools-jar-issue-with-jdk-11 -->
+            <groupId>com.nickwongdev</groupId>
+            <artifactId>aspectj-maven-plugin</artifactId>
+            <configuration>
+                <showWeaveInfo>false</showWeaveInfo>
+                <source>1.8</source>
+                <target>1.8</target>
+                <Xlint>ignore</Xlint>
+                <complianceLevel>1.8</complianceLevel>
+                <encoding>UTF-8</encoding>
+                <verbose>false</verbose>
+                <outxml>true</outxml>
+                <forceAjcCompile>true</forceAjcCompile>
+                <sources/><!-- this is important!-->
+                <aspectLibraries>
+                    <aspectLibrary>
+                        <groupId>com.microsoft.azure</groupId>
+                        <artifactId>azure-toolkit-common-lib</artifactId>
+                    </aspectLibrary>
+                </aspectLibraries>
+            </configuration>
+            <executions>
+                <execution>
+                    <id>compile-with-aspectj</id>
+                    <phase>process-classes</phase>
+                    <configuration>
+                        <weaveDirectories>
+                            <weaveDirectory>${project.build.directory}/classes</weaveDirectory>
+                        </weaveDirectories>
+                    </configuration>
+                    <goals>
+                        <goal>compile</goal>
+                    </goals>
+                </execution>
+                <execution>
+                    <id>test-compile-with-aspectj</id>
+                    <phase>process-test-classes</phase>
+                    <configuration>
+                        <weaveDirectories>
+                            <weaveDirectory>${project.build.directory}/test-classes</weaveDirectory>
+                        </weaveDirectories>
+                    </configuration>
+                    <goals>
+                        <goal>test-compile</goal>
+                    </goals>
+                </execution>
+            </executions>
+        </plugin>
+    </plugins>
+</build>
+<dependencies>
+    <dependency>
+        <groupId>org.projectlombok</groupId>
+        <artifactId>lombok</artifactId>
+        <scope>compile</scope>
+    </dependency>
+    <dependency>
+        <groupId>com.azure.resourcemanager</groupId>
+        <artifactId>azure-resourcemanager-loganalytics</artifactId>
+    </dependency>
+    <dependency>
+        <groupId>com.microsoft.azure</groupId>
+        <artifactId>azure-toolkit-auth-lib</artifactId>
+    </dependency>
+</dependencies>
+</project>
