+++ azure-toolkit-libs/azure-toolkit-monitor-lib/src/main/java/com/microsoft/azure/toolkit/lib/monitor/LogAnalyticsServiceWorkspaceSubscription.java
@@ -1,4 +1,4 @@
-package com.microsoft.azure.toolkit.lib.applicationinsights.workspace;
+package com.microsoft.azure.toolkit.lib.monitor;
 
 import com.azure.resourcemanager.loganalytics.LogAnalyticsManager;
 import com.microsoft.azure.toolkit.lib.common.model.AbstractAzResourceModule;
