+++ azure-toolkit-libs/azure-toolkit-monitor-lib/src/main/resources/META-INF/services/com.microsoft.azure.toolkit.lib.AzService
@@ -1,0 +1,1 @@
+com.microsoft.azure.toolkit.lib.monitor.AzureLogAnalyticsWorkspace
